Corrected bugs introduced due to merge
