+++ src/TextureWebAppChrome.js
@@ -1,36 +1,32 @@
 /* global vfs */
-import {
-  parseKeyEvent, VfsStorageClient, HttpStorageClient, InMemoryDarBuffer
-} from 'substance'
+import { parseKeyEvent } from 'substance'
+import DocumentArchiveFactory from './dar/DocumentArchiveFactory'
 import TextureAppChrome from './TextureAppChrome'
+import TextureArchive from './TextureArchive'
 
 export default class TextureWebAppChrome extends TextureAppChrome {
-  async _loadArchive (archiveId, context) {
-    let storage = this._getStorage(this.props.storageType)
-    let buffer = new InMemoryDarBuffer()
-    let ArchiveClass = this._getArchiveClass()
-    let archive = new ArchiveClass(storage, buffer, context)
+
+  async _loadArchive(archiveId, context) {
+    let documentArchiveConfig = this.props.documentArchiveConfig
+    documentArchiveConfig.setContext(context)
+    
+    let archive = DocumentArchiveFactory.getDocumentArchive(documentArchiveConfig)
+    
+    if (!archive)
+    {
+      archive = new TextureArchive(documentArchiveConfig)
+    }
+
     return archive.load(archiveId)
   }
 
-  _getStorage (storageType) {
-    if (storageType === 'vfs') {
-      return new VfsStorageClient(vfs, this._getDefaultDataFolder())
-    } else {
-      return new HttpStorageClient(this.props.storageUrl)
-    }
-  }
-
-  _handleKeyDown (event) {
+  _handleKeyDown(event) {
     let key = parseKeyEvent(event)
     // CommandOrControl+S
     if (key === 'META+83' || key === 'CTRL+83') {
+      console.log("Handling keydown event")
       this._save()
       event.preventDefault()
     }
   }
-
-  _getArchiveClass () { throw new Error('This method is abstract') }
-
-  _getDefaultDataFolder () { throw new Error('This method  is abstract') }
 }
