+++ src/index.js
@@ -10,9 +10,9 @@ export { default as Texture } from './Texture'
 export { default as TextureAppChrome } from './TextureAppChrome'
 export { default as TextureArchive } from './TextureArchive'
 export { default as TextureConfigurator } from './TextureConfigurator'
-export { default as TextureDesktopApp } from './TextureDesktopApp'
 export { default as TextureDesktopAppChrome } from './TextureDesktopAppChrome'
-export { default as TextureWebApp } from './TextureWebApp'
+export { default as TextureEditorWebApp } from './TextureEditorWebApp'
+export { default as TextureReaderWebApp } from './TextureReaderWebApp'
 export { default as TextureWebAppChrome } from './TextureWebAppChrome'
 
 export { default as textureGlobals } from './textureGlobals'
