+++ src/sessions/EditorSessionsGenerator.js
@@ -2,8 +2,8 @@ import { EditorSession, forEach } from "substance"
 
 import DocumentLoader from "../dar/DocumentLoader"
 import ManifestLoaderNew from "../dar/ManifestLoaderNew"
-import PubMetaLoader from "../PubMetaLoader"
-import TextureConfigurator from "../editor/util/TextureConfigurator"
+import PubMetaLoader from "../entities/PubMetaLoader"
+import TextureConfigurator from "../TextureConfigurator"
 
 /** 
  * @module editor/util/EditorSessionsGenerator
