+++ src/util/index.js
@@ -1,4 +1,3 @@
-export { default as checkArchive } from './checkArchive'
 export { default as createEmptyRef } from './createEmptyRef'
 export { default as removeElementAndXrefs } from './removeElementAndXrefs'
 export { default as updateEntityChildArray } from './updateEntityChildArray'
